Messaging Integrated With Game Control
I made some updates where I worked the game control subsystem and
messaging subsystem together. Now the client can choose to start a game
and this will trigger the game control system to assign characters to
the players. It will broadcast this information to all the clients. I
added logoff capability as well so this broadcasts. I wanted to do this
one as a model so we can go over the communication flow.