Updated GameBoard
Gameboard now includes the render board with support for player and room objects. It can now also return a list of possible moves. I have a place holder class for player while Andy works on player class, but we can hopefully get that integrated. I had some trouble using the package structure, so I removed it for the time being. If Sean or anyone could explain to me how we can compile and test with the package structure, that would be great. Stil have to do server.broadcast to render gameboard. Looking to start that tomorrow.